+++ templates/public.php
@@ -24,19 +24,6 @@ Util::addScript('analytics', '3rdParty/chartjs-plugin-datalabels.min');
 Util::addScript('analytics', '3rdParty/chartjs-plugin-zoom.min');
 ?>
 
-<header>
-    <div id="header">
-        <div id="header-left">
-            <a href="#"
-               title="" id="nextcloud">
-                <div class="logo logo-icon"></div>
-                <h1 class="header-appname" id="header-appname">
-                    Analytics for Nextcloud
-                </h1>
-            </a>
-        </div>
-    </div>
-</header>
 <div id="app-content">
     <?php print_unescaped($this->inc('part.content')); ?>
 </div>